Create Train speech-text model
This example uses the Kaldi speech recognition toolkit and specifies some example hyperparameters for training. Note that you will need to modify the code to use the specific speech-to-text model you want to train, and you'll also need to replace your-bucket-name and path/to with the correct values for your S3 bucket and file paths.